Update README.md (#40)
To be honest after reading the whole Readme, this is my indie app, but my income is not based on it as I have a separate job. Therefore I think it's unfair to get it promoted here when others are struggling with them.

However, I'll still help others to get promotion!